requirements.txt
alembic==1.13.2
beautifulsoup4==4.12.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pdfkit==1.0.0
python-dotenv==1.0.1
soupsieve==2.5
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
